Move logic to base
