+++ app/version_policy_providers/dockerhub.py
@@ -1,7 +1,6 @@
 from collections.abc import Iterator
 
 import requests
-from semantic_version import NpmSpec, Version
 
 from app.version_policy_providers.base import VersionPolicyProvider
 
@@ -33,35 +32,3 @@ class DockerhubVersionPolicyProvider(VersionPolicyProvider):
         for data in self.__get_all_api_pages():
             for result in data["results"]:
                 yield result["name"]
-
-    def get_all_semver_tags(
-        self, *, allow_prerelease: bool = False
-    ) -> Iterator[Version]:
-        for tag in self.get_all_tags():
-            try:
-                v = Version(tag)
-                if v.prerelease and not allow_prerelease:
-                    continue
-
-                yield v
-            except ValueError:
-                continue
-
-    def get_all_semver_tags_by_specifier(
-        self, specifier: str, *, allow_prerelease: bool = False
-    ) -> Iterator[Version]:
-        spec = NpmSpec(specifier)
-        return spec.filter(self.get_all_semver_tags(allow_prerelease=allow_prerelease))
-
-    def get_latest(
-        self, specifier: str, *, allow_prerelease: bool = False
-    ) -> Version | None:
-        try:
-            return max(
-                self.get_all_semver_tags_by_specifier(
-                    specifier, allow_prerelease=allow_prerelease
-                )
-            )
-        except ValueError:
-            # if `max` is called on an empty sequence
-            return None
